Add: 리뷰작성이 사진추가, 리뷰모델 개선

+++ app/src/main/java/com/cookandroid/gocafestudy/activities/ActivitySavedCafes.java
@@ -22,6 +22,7 @@ import com.cookandroid.gocafestudy.models.GET.Bookmark;
 import com.cookandroid.gocafestudy.R;
 import com.cookandroid.gocafestudy.models.GET.CafeDetail;
 import com.cookandroid.gocafestudy.models.GET.Review;
+import com.cookandroid.gocafestudy.models.POST.BookmarkCreateResponse;
 import com.cookandroid.gocafestudy.repository.MockRepository;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
 
@@ -92,12 +93,13 @@ public class ActivitySavedCafes extends AppCompatActivity {
         tvAiSummary.setText(cafe.getDescription());
         tvRating.setText(String.format("%.1f / 5.0", cafe.getReviewAverage()));
 
+        // --- 이미지 연결 ---
         List<String> images = cafe.getImages();
-        if (images.size() > 0) Glide.with(this).load(images.get(0)).into(ivMain);
-        if (images.size() > 1) Glide.with(this).load(images.get(1)).into(ivSub1);
-        if (images.size() > 2) Glide.with(this).load(images.get(2)).into(ivSub2);
-        if (images.size() > 3) Glide.with(this).load(images.get(3)).into(ivSub3);
-        if (images.size() > 4) Glide.with(this).load(images.get(4)).into(ivSub4);
+        if (images.size() > 0) Glide.with(this).load(images.get(0)).placeholder(R.drawable.ic_cafe1_img).into(ivMain);
+        if (images.size() > 1) Glide.with(this).load(images.get(1)).placeholder(R.drawable.ic_cafe1_img).into(ivSub1);
+        if (images.size() > 2) Glide.with(this).load(images.get(2)).placeholder(R.drawable.ic_cafe1_img).into(ivSub2);
+        if (images.size() > 3) Glide.with(this).load(images.get(3)).placeholder(R.drawable.ic_cafe1_img).into(ivSub3);
+        if (images.size() > 4) Glide.with(this).load(images.get(4)).placeholder(R.drawable.ic_cafe1_img).into(ivSub4);
 
         // 최근 리뷰 3개
         List<Review> recentReviews = cafe.getRecentReviews();
@@ -112,14 +114,21 @@ public class ActivitySavedCafes extends AppCompatActivity {
             dialog.dismiss();
         });
 
-        // 삭제 버튼
         Button btnSave = v.findViewById(R.id.btn_save);
-        btnSave.setText("삭제");
-        btnSave.setOnClickListener(click -> {
-            BookmarkDeleteResponse response = repository.deleteBookmark(cafeId);
-            Toast.makeText(this, response.message, Toast.LENGTH_SHORT).show();
-            dialog.dismiss();
-        });
+        if (cafe.isSaved()) {
+            // true일 때
+            btnSave.setText("삭제");
+            btnSave.setOnClickListener(click -> {
+                BookmarkDeleteResponse response = repository.deleteBookmark(cafeId);
+                Toast.makeText(this, response.message, Toast.LENGTH_SHORT).show();
+            });
+        } else {
+            // false일 때
+            btnSave.setOnClickListener(click -> {
+                BookmarkCreateResponse response = repository.createBookmark(cafeId);
+                Toast.makeText(this, response.message, Toast.LENGTH_SHORT).show();
+            });
+        }
 
         dialog.setContentView(v);
         dialog.show();
